+++ packages/data/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @apex-stack/data
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [26d83ca]
+  - @apex-stack/core@0.4.0
+
 ## 0.1.9
 
 ### Patch Changes