+++ playground/base-camp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # base-camp-spike
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [26d83ca]
+  - @apex-stack/core@0.4.0
+
 ## 0.0.5
 
 ### Patch Changes